Smaller changes due to compiler warnings
git-svn-id: https://subversion.gsi.de/fairroot/cbmroot/trunk@4782 0381ead4-6506-0410-b988-94b70fbc4730

+++ htrack/DataObjectLIB/include/specialMem.h
@@ -501,14 +501,14 @@ template<class objectClass> double specialMem<objectClass>::getAllocatedSizeOfDa
 	restorePointer = activeObjectPointer;
 
 	resetActiveObject();
-	for (int i = 0; i < getNumberOfEntries(); i++) {
+	for (unsigned long i = 0; i < getNumberOfEntries(); i++) {
 			
 		object       = readActiveObjectAndMakeNextOneActive();
-		returnValue += sizeof(object);
+		returnValue += (double)sizeof(object);
 
 	}
 
-	returnValue += (stackMem.capacity() - getNumberOfEntries()) * sizeof(objectClass);
+	returnValue += ((double)stackMem.capacity() - (double)getNumberOfEntries()) * (double)sizeof(objectClass);
 
 	activeObjectPointer = restorePointer;
 
@@ -534,16 +534,16 @@ template<class objectClass> double specialMem<objectClass>::getUsedSizeOfData(un
 	restorePointer = activeObjectPointer;
 
 	resetActiveObject();
-	for (int i = 0; i < getNumberOfEntries(); i++) {
+	for (unsigned long i = 0; i < getNumberOfEntries(); i++) {
 			
 		object       = readActiveObjectAndMakeNextOneActive();
-		returnValue += sizeof(object);
+		returnValue += (double)sizeof(object);
 
 	}
 
 	activeObjectPointer = restorePointer;
 
-	returnValue += sizeof(activeObjectPointer);
+	returnValue += (double)sizeof(activeObjectPointer);
 
 	returnValue  = (returnValue / (1 << (10 * dimension)));
